Move authentication tests and pass in db details
