+++ tests/routers/v1/application_test.py
@@ -15,6 +15,16 @@ from laa_crime_application_store_app.models.application_version_schema import (
 logger = structlog.getLogger(__name__)
 
 
+def test_applications_returns_401_when_authentication_required(authenticated_client):
+    response = authenticated_client.get("/v1/applications")
+    assert response.status_code == 401
+
+
+def test_applications_returns_200_when_unauthenticated(unauthenticated_client):
+    response = unauthenticated_client.get("/v1/applications")
+    assert response.status_code == 200
+
+
 def test_no_applications_return_empty_array(client: TestClient):
     response = client.get("/v1/applications")
     assert response.status_code == 200
